Move profile view to ImagerProfile app
