+++ imager/ImagerProfile/views.py
@@ -1,3 +1,36 @@
 from django.shortcuts import render
+from django.contrib.auth.decorators import login_required
+# from django.core.exceptions import ObjectDoesNotExist
+from imager_images.models import ImagerPhoto, ImagerAlbum, User
+from django.views.generic import UpdateView, DetailView
+from imagerprofile.models import ImagerProfile
 
-# Create your views here.
+
+class ImagerProfileDetailView(DetailView):
+    model = ImagerProfile
+
+
+class ImagerProfileUpdateView(UpdateView):
+    model = ImagerProfile
+    # form_class = ImagerProfileEditFeature
+
+
+@login_required()
+def profile(request):
+    # import pdb; pdb.set_trace()
+    user = request.user
+    try:
+        albums = ImagerAlbum.objects.filter(user=user).all()
+    except:
+        albums = None
+    profile = ImagerProfile.objects.get(user=user)
+    following = len(profile.followers.all())
+    num_photos = ImagerPhoto.objects.filter(user=user).count()
+    return render(request, 'profile.html', {
+        'albums': albums,
+        'user': user,
+        'profile': profile,
+        'following': following,
+        'num_album': len(albums),
+        'num_photo': num_photos
+    })
